Update VMwareWorkstationPlayer.json
The current version of this file returns version 16.2.5 for VMware Workstation Player instead of the expected version 17.0.1 (as of the time of writing). Although I do not have the skills and experience to implement a fix for this issue, I have found that the following API can be used to obtain the correct version information:

https://customerconnect.vmware.com/channel/public/api/v1.0/products/getProductHeader?locale=en_US&product=vmware_workstation_player&category=desktop_end_user_computing&version=14_0

This API returns the version ID "17_0" in JSON format, which can be used as the 'version' parameter for the 'getRelatedDLGList' function. For example:

https://customerconnect.vmware.com/channel/public/api/v1.0/products/getRelatedDLGList?&category=desktop_end_user_computing&product=vmware_workstation_player&dlgType=PRODUCT_BINARY&version=17_0

Call the above getRelatedDLGList API call will return the "code" (e.g., "WKST-PLAYER-1701") and "productId" (e.g., "1377") needed to complete the update URI.